Fix compiler warnings; whitespace

+++ src/hardware.c
@@ -73,7 +73,7 @@ void SysTick_Handler(void)
 {
     g_counter++;
 
-    for (int i = 0; i < N_BUTTONS; i++)
+    for (uint8_t i = 0; i < N_BUTTONS; i++)
     {
         const gpio_t *gpio = &g_buttons_gpio[i];
         button_t *button = &g_buttons[i];
@@ -107,7 +107,7 @@ void SysTick_Handler(void)
         }
     }
 
-    for (int i = 0; i < N_LEDS; i++)
+    for (uint8_t i = 0; i < N_LEDS; i++)
     {
         blinking_led_t *led = &g_blinking_led[i];
 
@@ -147,7 +147,7 @@ void SysTick_Handler(void)
 
                 led->time = 0;
             }
-        } 
+        }
     }
 }
 
@@ -206,7 +206,7 @@ void hw_init(void)
     delay_init();
 
     // leds
-    for (int i = 0; i < N_LEDS; i++)
+    for (uint8_t i = 0; i < N_LEDS; i++)
     {
         const gpio_t *gpio = &g_leds_gpio[i];
         Chip_GPIO_SetPinDIROutput(LPC_GPIO, gpio->port, gpio->pin);
@@ -214,7 +214,7 @@ void hw_init(void)
     }
 
     // buttons
-    for (int i = 0; i < N_BUTTONS; i++)
+    for (uint8_t i = 0; i < N_BUTTONS; i++)
     {
         const gpio_t *gpio = &g_buttons_gpio[i];
         Chip_GPIO_SetPinDIRInput(LPC_GPIO, gpio->port, gpio->pin);
@@ -228,7 +228,7 @@ void hw_init(void)
     clcd_init(CLCD_4BIT | CLCD_2LINE, &lcd2_gpio);
 
     // backlights
-    for (int i = 0; i < N_BACKLIGHTS; i++)
+    for (uint8_t i = 0; i < N_BACKLIGHTS; i++)
     {
         const gpio_t *gpio = &g_backlights_gpio[i];
         Chip_GPIO_SetPinDIROutput(LPC_GPIO, gpio->port, gpio->pin);
@@ -276,44 +276,44 @@ void hw_led_set(int led, int color, int value, int on_time_ms, int off_time_ms)
     if (color == LED_W)
     {
         colors[0] = LED_R;
-        colors[1] = LED_G;   
+        colors[1] = LED_G;
         colors[2] = LED_B;
     }
     else if (color == LED_Y)
     {
         colors[0] = LED_R;
         colors[1] = LED_G;
-        colors[2] = -1;   
+        colors[2] = -1;
     }
     else if(color == LED_C)
     {
         colors[0] = LED_B;
         colors[1] = LED_G;
-        colors[2] = -1;   
+        colors[2] = -1;
     }
     else if (color == LED_M)
     {
         colors[0] = LED_R;
         colors[1] = LED_B;
-        colors[2] = -1;   
+        colors[2] = -1;
     }
     else if (color == LED_R)
     {
         colors[0] = LED_R;
         colors[1] = -1;
-        colors[2] = -1;   
+        colors[2] = -1;
     }
     else if (color == LED_G)
     {
         colors[0] = -1;
         colors[1] = -1;
-        colors[2] = LED_G;   
+        colors[2] = LED_G;
     }
     else if (color == LED_B)
     {
         colors[0] = -1;
         colors[1] = LED_B;
-        colors[2] = -1;   
+        colors[2] = -1;
     }
 
     for (uint8_t j=0; j < 3; j++)
